api/requirements.txt
fastapi==0.104.1
supabase==2.0.2
python-dotenv==1.0.0
pydantic>=2.0.0